+++ estore-api/src/main/java/com/estore/api/estoreapi/persistence/UserFileDAO.java
@@ -256,7 +256,7 @@ public class UserFileDAO implements UserDAO {
    ** {@inheritDoc}
    */
   @Override
-  public Boolean AddToCart(String username, int beefId, float weight) throws IOException {
+  public Boolean AddToCart(String username, int beefId, double weight) throws IOException {
     synchronized (users) {
 
       Customer customer = GetCustomer(username);
